grc_backend/grc/jwt_auth.py
"""
Unified JWT Authentication for GRC and TPRM
This module provides a single, robust JWT authentication class that works for both GRC and TPRM modules.
"""
import jwt
import logging
from django.conf import settings
from django.contrib.auth import get_user_model
from rest_framework.authentication import BaseAuthentication
from rest_framework.exceptions import AuthenticationFailed

logger = logging.getLogger(__name__)


class UnifiedJWTAuthentication(BaseAuthentication):
    """
    Custom JWT authentication for DRF that authenticates users based on JWT tokens.
    Works for both GRC and TPRM modules with GRC user credentials.
    """
    
    def authenticate(self, request):
        """
        Authenticate the request and return a two-tuple of (user, token).
        """
        auth_header = request.headers.get('Authorization')
        
        if not auth_header:
            return None
        
        if not auth_header.startswith('Bearer '):
            logger.warning("[Unified JWT Auth] Invalid authentication header format. Expected: Bearer <token>")
            return None
        
        token = auth_header.split(' ')[1]
        
        try:
            # Decode JWT token
            secret_key = getattr(settings, 'JWT_SECRET_KEY', settings.SECRET_KEY)
            payload = jwt.decode(token, secret_key, algorithms=['HS256'])
            
            user_id = payload.get('user_id')
            username = payload.get('username')
            
            logger.info(f"[Unified JWT Auth] Token decoded successfully, user_id: {user_id}, username: {username}")
            
            if not user_id:
                logger.warning("[Unified JWT Auth] Token does not contain user_id")
                raise AuthenticationFailed('Token does not contain user_id')
            
            # Try to get the user from the database
            User = get_user_model()
            
            try:
                user = User.objects.get(pk=user_id)
                
                # Ensure the user object has is_authenticated property
                if not hasattr(user, 'is_authenticated'):
                    user.is_authenticated = True
                
                # Add userid for TPRM compatibility
                if not hasattr(user, 'userid'):
                    user.userid = user.pk
                
                logger.info(f"[Unified JWT Auth] GRC User authenticated: {user.username}")
                return (user, token)
                
            except User.DoesNotExist:
                logger.warning(f"[Unified JWT Auth] User with ID {user_id} not found in database. Creating MockUser.")
                
                # Create a mock user for cases where user might not exist in local DB
                class MockUser:
                    def __init__(self, user_id, username):
                        self.pk = user_id
                        self.id = user_id
                        self.userid = user_id
                        self.username = username if username else f"user_{user_id}"
                        self.is_authenticated = True
                        self.is_active = True
                        self.is_staff = False
                        self.is_superuser = False
                    
                    def __str__(self):
                        return self.username
                    
                    def get_full_name(self):
                        return self.username
                    
                    def get_short_name(self):
                        return self.username
                
                mock_user = MockUser(user_id, username)
                logger.info(f"[Unified JWT Auth] MockUser created: {mock_user.username}")
                return (mock_user, token)
                
            except Exception as db_error:
                logger.error(f"[Unified JWT Auth] Database error during user lookup: {db_error}")
                
                # For database errors, still create a MockUser to allow access
                class MockUser:
                    def __init__(self, user_id, username):
                        self.pk = user_id
                        self.id = user_id
                        self.userid = user_id
                        self.username = username if username else f"user_{user_id}"
                        self.is_authenticated = True
                        self.is_active = True
                        self.is_staff = False
                        self.is_superuser = False
                    
                    def __str__(self):
                        return self.username
                    
                    def get_full_name(self):
                        return self.username
                    
                    def get_short_name(self):
                        return self.username
                
                mock_user = MockUser(user_id, username)
                logger.warning(f"[Unified JWT Auth] Database error, using MockUser: {mock_user.username}")
                return (mock_user, token)
        
        except jwt.ExpiredSignatureError:
            logger.warning("[Unified JWT Auth] JWT token expired")
            raise AuthenticationFailed('Token expired')
        
        except jwt.InvalidTokenError as e:
            logger.warning(f"[Unified JWT Auth] Invalid JWT token: {e}")
            raise AuthenticationFailed('Invalid token')
        
        except Exception as e:
            logger.error(f"[Unified JWT Auth] Unexpected error during JWT authentication: {e}")
            raise AuthenticationFailed(f"Authentication error: {e}")
