grc_backend/requirements.txt
Django>=4.2.7,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
PyJWT>=2.8.0
python-dotenv>=1.0.0
openpyxl>=3.1.2  # For Excel file handling (reading and writing)
xlsxwriter>=3.1.0  # For Excel file handling with advanced formatting
mysqlclient>=2.2.0
pymysql>=1.1.0
boto3>=1.34.0
mysql-connector-python>=8.0.0
pandas>=2.1.4
xmltodict
reportlab
requests>=2.31.0
celery>=5.3.4
python-docx
langchain_openai
langchain
langchain-core
langchain-community
openai
langchain_ollama
werkzeug
pdfplumber
PyPDF2
PyMuPDF>=1.23.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# =========================================================================
# TPRM INTEGRATION DEPENDENCIES
# =========================================================================
# Django extensions for TPRM
django-filter>=23.4
drf-yasg>=1.21.7
django-extensions>=3.2.3
django-import-export>=3.3.4
django-simple-history>=3.4.0

# Security packages for TPRM
python-decouple>=3.8
cryptography>=41.0.7
bcrypt>=4.1.2

# File Processing & OCR for TPRM
Pillow>=10.1.0
pytesseract>=0.3.10

# Data Processing
numpy>=1.25.2

# Storage & AWS
django-storages>=1.14.2

# Async & HTTP
aiohttp>=3.9.1

# Logging
python-json-logger>=2.0.7
django-debug-toolbar>=4.2.0

# SQLAlchemy for vendor module
SQLAlchemy>=2.0.23

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0